Create more packages including command, task, design, and exception Move some design to the design package Change the class name Design to Default (so that it only prints default design) Move all the boolean method for checking valid command to command package Move all the classes for superclass tasks and its subclasses to package task
