+++ src/main/java/duke/design/Default.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.design;
+
+import duke.task.Task;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -8,8 +10,13 @@ import java.util.Random;
  * Credit: The pictures are generated with the help of an online picture to ASCII symbols converter
  * The website available: "https://manytools.org/hacker-tools/convert-images-to-ascii-art/"
  */
-public abstract class Design {
+public abstract class Default {
     public static final String CURR_VERSION = "Version 5.0";
+    //Corner symbols for to-do list frames
+    public static final String TOP_LEFT_CORNER = "/";
+    public static final String TOP_RIGHT_CORNER = "\\";
+    public static final String BOTTOM_LEFT_CORNER = "\\";
+    public static final String BOTTOM_RIGHT_CORNER = "/";
 
     /**
      * Print the logo and greeting message
@@ -231,13 +238,13 @@ public abstract class Design {
         int randomNumber = random.nextInt(3);
         switch (randomNumber) {
         case 0:
-            System.out.println(Design.PICTURE_1);
+            System.out.println(Default.PICTURE_1);
             break;
         case 1:
-            System.out.println(Design.PICTURE_2);
+            System.out.println(Default.PICTURE_2);
             break;
         case 2:
-            System.out.println(Design.PICTURE_3);
+            System.out.println(Default.PICTURE_3);
             break;
         }
         System.out.println("**                                                                                                                   **");
@@ -284,5 +291,100 @@ public abstract class Design {
         //Exit command
         helpLists.add("To exit the program, use the command \"exit\" or \"bye\"\n");
         return helpLists;
+    }
+
+    /**
+     * Prints the to-do list with frames
+     *
+     * @param tasks     the array of class Task instance which stores all the tasks added by the user
+     * @param totalTask the last index of the array that is not null
+     */
+    public static void printToDoList(Task[] tasks, int totalTask, int longestTaskDescription) {
+        final int MIN_LENGTH = " My to-do list: ".length();
+        //if longestTaskDescription is shorter than the length of the string "My to-do list: ", sets it to the length of the string
+        if (longestTaskDescription < MIN_LENGTH) {
+            longestTaskDescription = MIN_LENGTH;
+        }
+        //Prints the to-do list
+        drawUpperFrame(longestTaskDescription);
+        printTasks(tasks, totalTask, longestTaskDescription);
+        drawLowerFrame(longestTaskDescription);
+    }
+
+    /**
+     * Prints the bottom frame of the to-do list and the guide for reading the to-do list
+     *
+     * @param longestTaskDescription the length of the longest task description string stored in the tasks array
+     */
+    private static void drawLowerFrame(int longestTaskDescription) {
+        System.out.print("\t" + BOTTOM_LEFT_CORNER);
+        for (int i = 0; i < longestTaskDescription + "| [ ][ ] 100. ".length(); i++) {
+            System.out.print("-");
+        }
+        System.out.println(BOTTOM_RIGHT_CORNER);
+        //Shows the guide for understanding the to-do list
+        System.out.println("\tFor your knowledge, ");
+        System.out.println("\tthe first [ ] indicates the type of the task ('T' for to-do, 'D' for deadline, 'E' for event)");
+        System.out.println("\tthe second [ ] indicates whether the task is completed:\n" +
+                "\t[X] when the task is marked completed\t[ ] when the task is not done.");
+    }
+
+    /**
+     * Prints the tasks stored in the array, the frame starts with '|' and ends with '|', the ending frame is always located at the position of the longest task description
+     * @param tasks the array that stores all the tasks
+     * @param totalTask the number of the tasks stored
+     * @param longestTaskDescription the length of the longest task description stored in the tasks array
+     */
+    private static void printTasks(Task[] tasks, int totalTask, int longestTaskDescription) {
+        for (int i = 0; i < totalTask; i++) {
+            //Fill the first [] with class type, and the second [] with a 'X' if the task is completed
+            if (tasks[i].getDone()) {
+                System.out.print("\t| [" + tasks[i].getClassType() + "][X] " + (i + 1) + ". ");
+            } else {
+                System.out.print("\t| [" + tasks[i].getClassType() + "][ ] " + (i + 1) + ". ");
+            }
+            //Calculates the required spacing for the current task as compared to the longest task description to print '|'
+            int distanceToClosingFrame = longestTaskDescription + "| [ ][ ] 100. ".length() - ("| [ ][ ] " + (i + 1) + ". ").length() + 1;
+            System.out.printf("%1$-" + distanceToClosingFrame + "s", tasks[i]);
+            System.out.println("|");
+        }
+    }
+
+    /**
+     * Prints the upper frame of the to-do list and its default display string
+     *
+     * @param longestTaskDescription the length of the longest task description string stored in the tasks array
+     */
+    private static void drawUpperFrame(int longestTaskDescription) {
+        System.out.print("\t" + TOP_LEFT_CORNER); //the top left corner
+        for (int i = 0; i < longestTaskDescription + "| [ ][ ] 100. ".length(); i++) {
+            System.out.print("-");
+        }
+        System.out.println(TOP_RIGHT_CORNER);
+        //Print default string " My to-do list: "
+        System.out.print("\t| My to-do list: ");
+        for (int i = 0; i < longestTaskDescription + "| [ ][ ] 100. ".length() - "| My to-do list: ".length() + 1; i++) {
+            System.out.print(" ");
+        }
+        System.out.println("|");
+    }
+
+    /**
+     * Shows the formatted message string
+     *
+     * @param message The message to print
+     */
+    public static void showMessage(String message) {
+        System.out.print("\t@");
+        for (int i = 0; i < message.length() + 4; i++) {
+            System.out.print("-");
+        }
+        System.out.println("@");
+        System.out.println("\t   " + message);
+        System.out.print("\t@");
+        for (int i = 0; i < message.length() + 4; i++) {
+            System.out.print("-");
+        }
+        System.out.println("@");
     }
 }
