+++ src/main/java/duke/exception/DukeException.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.exception;
 
 public class DukeException extends Exception {
     public class InvalidCommandException{
